@@ -1,6 +1,4 @@
 ﻿using ICSharpCode.SharpZipLib.GZip;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using ProtoBuf;
 using System;
 using System.Collections.Generic;
@@ -11,6 +9,7 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
+using System.Text.Json;
 using static BBDown.BBDownLogger;
 
 namespace BBDown
@@ -255,7 +254,7 @@ namespace BBDown
         /// <returns></returns>
         private static string ConvertToString(object obj)
         {
-            return JsonConvert.SerializeObject(obj);
+            return JsonSerializer.Serialize(obj, obj.GetType());
         }
 
         /// <summary>
